Add vitest and operator utilities tests
